LayoutTests/storage/indexeddb/index-basics.html

<html>
<head>
<link rel="stylesheet" href="../../fast/js/resources/js-test-style.css">
<script src="../../fast/js/resources/js-test-pre.js"></script>
<script src="../../fast/js/resources/js-test-post-function.js"></script>
<script src="resources/shared.js"></script>
</head>
<body>
<p id="description"></p>
<div id="console"></div>
<script>

description("Test the basics of IndexedDB's webkitIDBIndex.");
if (window.layoutTestController) 
    layoutTestController.waitUntilDone();

function test()
{
    result = evalAndLog("webkitIndexedDB.open('name')");
    verifyResult(result);
    result.onsuccess = setVersion;
    result.onerror = unexpectedErrorCallback;
}

function setVersion()
{
    verifySuccessEvent(event);
    db = evalAndLog("db = event.result");

    result = evalAndLog("db.setVersion('new version')");
    verifyResult(result);
    result.onsuccess = deleteExisting;
    result.onerror = unexpectedErrorCallback;
}

function deleteExisting()
{
    debug("setVersionSuccess():");
    verifySuccessEvent(event);
    window.trans = evalAndLog("trans = event.result");
    shouldBeTrue("trans !== null");
    trans.onabort = unexpectedAbortCallback;

    deleteAllObjectStores(db, createIndex);
}

function createIndex()
{
    window.store = evalAndLog("db.createObjectStore('storeName', null)");
    window.indexObject = evalAndLog("store.createIndex('indexName', 'x')");
    window.indexObject2 = evalAndLog("store.createIndex('indexName2', 'y', false)");
    window.indexObject3 = evalAndLog("store.createIndex('zIndex', 'z', true)");
    addData();
}

function addData()
{
    shouldBeTrue("'name' in indexObject");
    shouldBeEqualToString("indexObject.name", "indexName");
    shouldBeTrue("'storeName' in indexObject");
    shouldBeEqualToString("indexObject.storeName", "storeName");
    shouldBeTrue("'keyPath' in indexObject");
    shouldBeEqualToString("indexObject.keyPath", "x");
    shouldBeTrue("'unique' in indexObject");
    shouldBeFalse("indexObject.unique");
    shouldBeTrue("'openKeyCursor' in indexObject");
    shouldBeTrue("'openCursor' in indexObject");
    shouldBeTrue("'getKey' in indexObject");
    shouldBeTrue("'get' in indexObject");

    result = evalAndLog("store.add({x: 'value', y: 'zzz', z: 2.72}, 'key')");
    verifyResult(result);
    result.onsuccess = addMore;
    result.onerror = unexpectedErrorCallback;
}

function addMore()
{
    verifySuccessEvent(event);

    result = evalAndLog("event.source.add({x: 'value2', y: 'zzz2', z: 2.71}, 'key2')");
    verifyResult(result);
    result.onsuccess = getData;
    result.onerror = unexpectedErrorCallback;
}

function getData()
{
    verifySuccessEvent(event);

    result = evalAndLog("indexObject.getKey('value')");
    verifyResult(result);
    result.onsuccess = getObjectData;
    result.onerror = unexpectedErrorCallback;
}

function getObjectData()
{
    verifySuccessEvent(event);
    shouldBeEqualToString("event.result", "key");

    result = evalAndLog("indexObject2.getKey('zzz')");
    verifyResult(result);
    result.onsuccess = getObjectData2;
    result.onerror = unexpectedErrorCallback;
}

function getObjectData2()
{
    verifySuccessEvent(event);
    shouldBeEqualToString("event.result", "key");

    result = evalAndLog("indexObject3.get(2.71)");
    verifyResult(result);
    result.onsuccess = getObjectData3;
    result.onerror = unexpectedErrorCallback;
}

function getObjectData3()
{
    verifySuccessEvent(event);
    shouldBeEqualToString("event.result.x", "value2");

    result = evalAndLog("indexObject.get('value')");
    verifyResult(result);
    result.onsuccess = getDataFail;
    result.onerror = unexpectedErrorCallback;
}

function getDataFail()
{
    verifySuccessEvent(event);
    shouldBeEqualToString("event.result.x", "value");
    shouldBeEqualToString("event.result.y", "zzz");

    result = evalAndLog("indexObject.getKey('does not exist')");
    verifyResult(result);
    result.onsuccess = unexpectedSuccessCallback;
    result.onerror = getObjectDataFail;
}

function getObjectDataFail()
{
    verifyErrorEvent(event);
    shouldBe("event.code", "webkitIDBDatabaseException.NOT_FOUND_ERR");

    result = evalAndLog("indexObject.get('does not exist')");
    verifyResult(result);
    result.onsuccess = unexpectedSuccessCallback;
    result.onerror = openKeyCursor;
}

function openKeyCursor()
{
    verifyErrorEvent(event);
    shouldBe("event.code", "webkitIDBDatabaseException.NOT_FOUND_ERR");

    window.result = evalAndLog("indexObject.openKeyCursor()");
    verifyResult(result);
    result.onsuccess = cursor1Continue;
    result.onerror = unexpectedErrorCallback;
}

function cursor1Continue()
{
    verifySuccessEvent(event);
    shouldBeFalse("event.result === null");
    shouldBeEqualToString("event.result.key", "value");
    shouldBeEqualToString("event.result.value", "key");

    // We re-use the last result object.
    evalAndLog("event.result.continue()");
    verifyResult(window.result);
    window.result.onsuccess = cursor1Continue2;
}

function cursor1Continue2()
{
    verifySuccessEvent(event);
    shouldBeFalse("event.result === null");
    shouldBeEqualToString("event.result.key", "value2");
    shouldBeEqualToString("event.result.value", "key2");

    // We re-use the last result object.
    evalAndLog("event.result.continue()");
    verifyResult(window.result);
    window.result.onsuccess = openObjectCursor;
}

function openObjectCursor()
{
    verifySuccessEvent(event);
    shouldBeTrue("event.result === null");

    window.result = evalAndLog("indexObject.openCursor()");
    verifyResult(result);
    result.onsuccess = cursor2Continue;
    result.onerror = unexpectedErrorCallback;
}

function cursor2Continue()
{
    verifySuccessEvent(event);
    shouldBeFalse("event.result === null");
    shouldBeEqualToString("event.result.key", "value");
    shouldBeEqualToString("event.result.value.x", "value");
    shouldBeEqualToString("event.result.value.y", "zzz");

    // We re-use the last result object.
    evalAndLog("event.result.continue()");
    verifyResult(window.result);
    window.result.onsuccess = cursor2Continue2;
}

function cursor2Continue2()
{
    verifySuccessEvent(event);
    shouldBeFalse("event.result === null");
    shouldBeEqualToString("event.result.key", "value2");
    shouldBeEqualToString("event.result.value.x", "value2");    
    shouldBeEqualToString("event.result.value.y", "zzz2");

    // We re-use the last result object.
    evalAndLog("event.result.continue()");
    verifyResult(window.result);
    window.result.onsuccess = last;
}

function last()
{
    verifySuccessEvent(event);
    shouldBeTrue("event.result === null");

    try {
        debug("Passing an invalid key into indexObject.get().");
        indexObject.get([]);
        testFailed("No exception thrown");
    } catch (e) {
        testPassed("Caught exception: " + e.toString());
    }

    try {
        debug("Passing an invalid key into indexObject.getKey().");
        indexObject.getKey([]);
        testFailed("No exception thrown");
    } catch (e) {
        testPassed("Caught exception: " + e.toString());
    }
    done();
}

test();

var successfullyParsed = true;

</script>
</body>
</html>
